Refactor code
Export default MAP function in createMapReducer into index.js
