+++ src/index.js
@@ -6,7 +6,7 @@ import locationsReducer from './store/reducers/locations';
 import regions from './store/reducers/regions';
 import styles from './store/reducers/styles';
 import supersetConfig from './store/reducers/superset-config';
-import MAP from './store/reducers/createMapReducer';
+import MAP, { createMapReducer } from './store/reducers/map';
 
 // eslint-disable-next-line no-underscore-dangle, global-require
 if (!global._babelPolyfill) require('babel-polyfill');
@@ -17,8 +17,8 @@ gisida.version = require('../package.json').version;
 gisida.initStore = require('./store/initStore').default;
 gisida.loadLayers = require('./store/initStore').loadLayers;
 gisida.reducerRegistry = require('./store/reducerRegistry').default;
-gisida.createMapReducer = require('./store/reducers/createMapReducer/createMapReducer').createMapReducer;
 
+gisida.createMapReducer = createMapReducer
 gisida.defaultReducers = {
   APP,
   filter,
