--- src/store/reducers/createMapReducer/index.js
@@ -1,1 +1,0 @@
-export default './createMapReducer';
