+++ src/store/reducers/map/activeLayerIdsReducer.js
