+++ src/store/reducers/map/createMapReducer.js
