+++ src/store/reducers/map/currentRegionReducer.js
