+++ src/store/reducers/map/currentStyleReducer.js
