+++ src/store/reducers/map/defaultLayersReducer.js
