+++ src/store/reducers/map/detailViewReducer.js
