+++ src/store/reducers/map/filterReducer.js
