+++ src/store/reducers/map/index.js
@@ -1,0 +1,4 @@
+import MAP, { createMapReducer } from './createMapReducer';
+
+export { createMapReducer }; // exports the function to create a Map reducer
+export default MAP; // exports the default Map reducer as the default export
