+++ src/store/reducers/map/layerVisibilityReducer.js
