+++ src/store/reducers/map/layersReducer.js
