+++ src/store/reducers/map/mapLoadedReducer.js
