+++ src/store/reducers/map/mapRenderedReducer.js
