+++ src/store/reducers/map/menuIsOpenReducer.js
