+++ src/store/reducers/map/primaryLayerReducer.js
