+++ src/store/reducers/map/reloadLayersReducer.js
