+++ src/store/reducers/map/showFilterPanelReducer.js
