+++ src/store/reducers/map/showSpinnerReducer.js
