+++ src/store/reducers/map/tests/activeLayerIdReducer.test.js
