+++ src/store/reducers/map/tests/activeLayerIdsReducer.test.js
