+++ src/store/reducers/map/tests/common.js
