+++ src/store/reducers/map/tests/currentRegionReducer.test.js
