+++ src/store/reducers/map/tests/currentStyleReducer.test.js
