+++ src/store/reducers/map/tests/defaultLayersReducer.test.js
