+++ src/store/reducers/map/tests/detailViewReducer.test.js
