+++ src/store/reducers/map/tests/filterReducer.test.js
