+++ src/store/reducers/map/tests/lastLayerSelectedReducer.test.js
