+++ src/store/reducers/map/tests/layerVisibilityReducer.test.js
