+++ src/store/reducers/map/tests/layersReducer.test.js
