+++ src/store/reducers/map/tests/mapLoadedReducer.test.js
