+++ src/store/reducers/map/tests/mapRenderedReducer.test.js
