+++ src/store/reducers/map/tests/menuIsOpenReducer.test.js
