+++ src/store/reducers/map/tests/oldLayerObjsReducer.test.js
