+++ src/store/reducers/map/tests/primaryLayerReducer.test.js
