+++ src/store/reducers/map/tests/primarySubLayerReducer.test.js
