+++ src/store/reducers/map/tests/reloadLayerIdReducer.test.js
