+++ src/store/reducers/map/tests/reloadLayersReducer.test.js
