+++ src/store/reducers/map/tests/showFilterPanelReducer.test.js
