+++ src/store/reducers/map/tests/showSpinnerReducer.test.js
