+++ src/store/reducers/map/tests/timeSeriesReducer.test.js
