+++ src/store/reducers/map/timeSeriesReducer.js
